2通りのtodoをホーム画面に追加
