+++ lib/main.dart
@@ -1,72 +1,10 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:todos/firebase.dart';
+import 'package:todos/page/from_firestore_page.dart';
+import 'package:todos/page/basic_todo_page.dart';
 
-import 'firebase_options.dart';
-import 'todo.dart';
-
-/// Some keys used for testing
-final addTodoKey = UniqueKey();
-final activeFilterKey = UniqueKey();
-final completedFilterKey = UniqueKey();
-final allFilterKey = UniqueKey();
-
-/// Creates a [TodoList] and initialise it with pre-defined values.
-///
-/// We are using [StateNotifierProvider] here as a `List<Todo>` is a complex
-/// object, with advanced business logic like how to edit a todo.
-final todoListProvider = StateNotifierProvider<TodoList, List<Todo>>((ref) {
-  return TodoList([
-    Todo(todoId: 'todo-0', title: 'hi', createdAt: DateTime.now()),
-    Todo(todoId: 'todo-1', title: 'hello', createdAt: DateTime.now()),
-    Todo(todoId: 'todo-2', title: 'bonjour', createdAt: DateTime.now()),
-  ]);
-});
-
-/// The different ways to filter the list of todos
-enum TodoListFilter {
-  all,
-  active,
-  isCompleted,
-}
-
-/// The currently active filter.
-///
-/// We use [StateProvider] here as there is no fancy logic behind manipulating
-/// the value since it's just enum.
-final todoListFilter = StateProvider((_) => TodoListFilter.all);
-
-/// The number of uncompleted todos
-///
-/// By using [Provider], this value is cached, making it performant.\
-/// Even multiple widgets try to read the number of uncompleted todos,
-/// the value will be computed only once (until the todo-list changes).
-///
-/// This will also optimise unneeded rebuilds if the todo-list changes, but the
-/// number of uncompleted todos doesn't (such as when editing a todo).
-final uncompletedTodosCount = Provider<int>((ref) {
-  return ref.watch(todoListProvider).where((todo) => !todo.isCompleted).length;
-});
-
-/// The list of todos after applying of [todoListFilter].
-///
-/// This too uses [Provider], to avoid recomputing the filtered list unless either
-/// the filter of or the todo-list updates.
-final filteredTodos = Provider<List<Todo>>((ref) {
-  final filter = ref.watch(todoListFilter);
-  final todos = ref.watch(todoListProvider);
-
-  switch (filter) {
-    case TodoListFilter.isCompleted:
-      return todos.where((todo) => todo.isCompleted).toList();
-    case TodoListFilter.active:
-      return todos.where((todo) => !todo.isCompleted).toList();
-    case TodoListFilter.all:
-      return todos;
-  }
-});
+import 'firebase/firebase_options.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -80,242 +18,45 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: getFireStorePage(),
+      home: const Home(),
     );
   }
 }
 
-class Home extends HookConsumerWidget {
+class Home extends StatelessWidget {
   const Home({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final todos = ref.watch(filteredTodos);
-    final newTodoController = useTextEditingController();
-
-    return GestureDetector(
-      onTap: () => FocusScope.of(context).unfocus(),
-      child: Scaffold(
-        body: ListView(
-          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 40),
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('ホーム'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            const Title(),
-            TextField(
-              key: addTodoKey,
-              controller: newTodoController,
-              decoration: const InputDecoration(
-                labelText: 'What needs to be done?',
-              ),
-              onSubmitted: (value) {
-                ref.read(todoListProvider.notifier).add(value);
-                newTodoController.clear();
+            ElevatedButton(
+              onPressed: () {
+                Navigator.of(context)
+                    .push(MaterialPageRoute(builder: (context) {
+                  return BasicTodoPage();
+                }));
               },
+              child: const Text('ベーシック'),
             ),
-            const SizedBox(height: 42),
-            const Toolbar(),
-            if (todos.isNotEmpty) const Divider(height: 0),
-            for (var i = 0; i < todos.length; i++) ...[
-              if (i > 0) const Divider(height: 0),
-              Dismissible(
-                background: Container(
-                  color: Colors.red,
-                  padding: EdgeInsets.only(
-                    right: 10,
-                  ),
-                  alignment: AlignmentDirectional.centerEnd,
-                  child: Icon(
-                    Icons.delete,
-                    color: Colors.white,
-                  ),
-                ),
-                direction: DismissDirection.endToStart,
-                key: ValueKey(todos[i].todoId),
-                onDismissed: (direction) {
-                  ref.read(todoListProvider.notifier).remove(todos[i]);
-                },
-                child: ProviderScope(
-                  overrides: [
-                    _currentTodo.overrideWithValue(todos[i]),
-                  ],
-                  child: const TodoItem(),
-                ),
-              )
-            ],
+            ElevatedButton(
+              onPressed: () {
+                Navigator.of(context)
+                    .push(MaterialPageRoute(builder: (context) {
+                  return FromFireStorePage();
+                }));
+              },
+              child: const Text('from FireStore'),
+            ),
           ],
         ),
       ),
     );
   }
-}
-
-class Toolbar extends HookConsumerWidget {
-  const Toolbar({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final filter = ref.watch(todoListFilter);
-
-    Color? textColorFor(TodoListFilter value) {
-      return filter == value ? Colors.blue : Colors.black;
-    }
-
-    return Material(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Expanded(
-            child: Text(
-              '${ref.watch(uncompletedTodosCount).toString()} items left',
-              overflow: TextOverflow.ellipsis,
-            ),
-          ),
-          Tooltip(
-            key: allFilterKey,
-            message: 'All todos',
-            child: TextButton(
-              onPressed: () =>
-                  ref.read(todoListFilter.notifier).state = TodoListFilter.all,
-              style: ButtonStyle(
-                visualDensity: VisualDensity.compact,
-                foregroundColor:
-                    MaterialStateProperty.all(textColorFor(TodoListFilter.all)),
-              ),
-              child: const Text('All'),
-            ),
-          ),
-          Tooltip(
-            key: activeFilterKey,
-            message: 'Only uncompleted todos',
-            child: TextButton(
-              onPressed: () => ref.read(todoListFilter.notifier).state =
-                  TodoListFilter.active,
-              style: ButtonStyle(
-                visualDensity: VisualDensity.compact,
-                foregroundColor: MaterialStateProperty.all(
-                  textColorFor(TodoListFilter.active),
-                ),
-              ),
-              child: const Text('Active'),
-            ),
-          ),
-          Tooltip(
-            key: completedFilterKey,
-            message: 'Only completed todos',
-            child: TextButton(
-              onPressed: () => ref.read(todoListFilter.notifier).state =
-                  TodoListFilter.isCompleted,
-              style: ButtonStyle(
-                visualDensity: VisualDensity.compact,
-                foregroundColor: MaterialStateProperty.all(
-                  textColorFor(TodoListFilter.isCompleted),
-                ),
-              ),
-              child: const Text('Completed'),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class Title extends StatelessWidget {
-  const Title({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Text(
-      'todos',
-      textAlign: TextAlign.center,
-      style: TextStyle(
-        color: Color.fromARGB(38, 47, 47, 247),
-        fontSize: 100,
-        fontWeight: FontWeight.w100,
-        fontFamily: 'Helvetica Neue',
-      ),
-    );
-  }
-}
-
-/// A provider which exposes the [Todo] displayed by a [TodoItem].
-///
-/// By retrieving the [Todo] through a provider instead of through its
-/// constructor, this allows [TodoItem] to be instantiated using the `const` keyword.
-///
-/// This ensures that when we add/remove/edit todos, only what the
-/// impacted widgets rebuilds, instead of the entire list of items.
-final _currentTodo = Provider<Todo>((ref) => throw UnimplementedError());
-
-class TodoItem extends HookConsumerWidget {
-  const TodoItem({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final todo = ref.watch(_currentTodo);
-    final itemFocusNode = useFocusNode();
-    final itemIsFocused = useIsFocused(itemFocusNode);
-
-    final textEditingController = useTextEditingController();
-    final textFieldFocusNode = useFocusNode();
-
-    return Material(
-      color: Colors.white,
-      elevation: 6,
-      child: Focus(
-        focusNode: itemFocusNode,
-        onFocusChange: (focused) {
-          if (focused) {
-            textEditingController.text = todo.title;
-          } else {
-            // Commit changes only when the textfield is unfocused, for performance
-            ref
-                .read(todoListProvider.notifier)
-                .edit(todoId: todo.todoId, title: textEditingController.text);
-          }
-        },
-        child: ListTile(
-          onTap: () {
-            itemFocusNode.requestFocus();
-            textFieldFocusNode.requestFocus();
-          },
-          leading: Checkbox(
-            value: todo.isCompleted,
-            onChanged: (value) =>
-                ref.read(todoListProvider.notifier).toggle(todo.todoId),
-          ),
-          title: itemIsFocused
-              ? TextField(
-                  autofocus: true,
-                  focusNode: textFieldFocusNode,
-                  controller: textEditingController,
-                )
-              : Text(todo.title),
-          subtitle: Text(
-            '${todo.createdAt.year.toString()}年' +
-                '${todo.createdAt.month.toString()}月' +
-                '${todo.createdAt.day.toString()}日 ' +
-                '${todo.createdAt.hour.toString()}:' +
-                '${todo.createdAt.minute.toString()}',
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-bool useIsFocused(FocusNode node) {
-  final isFocused = useState(node.hasFocus);
-
-  useEffect(() {
-    void listener() {
-      isFocused.value = node.hasFocus;
-    }
-
-    node.addListener(listener);
-    return () => node.removeListener(listener);
-  }, [node]);
-
-  return isFocused.value;
 }
