+++ lib/page/basic_todo_page.dart
@@ -1,0 +1,282 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import '../data/todo.dart';
+
+
+final addTodoKey = UniqueKey();
+final activeFilterKey = UniqueKey();
+final completedFilterKey = UniqueKey();
+final allFilterKey = UniqueKey();
+
+final todoListProvider = StateNotifierProvider<TodoList, List<Todo>>((ref) {
+  return TodoList([
+    Todo(todoId: 'todo-0', title: 'hi', createdAt: DateTime.now()),
+    Todo(todoId: 'todo-1', title: 'hello', createdAt: DateTime.now()),
+    Todo(todoId: 'todo-2', title: 'bonjour', createdAt: DateTime.now()),
+  ]);
+});
+
+enum TodoListFilter {
+  all,
+  active,
+  isCompleted,
+}
+final todoListFilter = StateProvider((_) => TodoListFilter.all);
+
+final uncompletedTodosCount = Provider<int>((ref) {
+  return ref.watch(todoListProvider).where((todo) => !todo.isCompleted).length;
+});
+final filteredTodos = Provider<List<Todo>>((ref) {
+  final filter = ref.watch(todoListFilter);
+  final todos = ref.watch(todoListProvider);
+
+  switch (filter) {
+    case TodoListFilter.isCompleted:
+      return todos.where((todo) => todo.isCompleted).toList();
+    case TodoListFilter.active:
+      return todos.where((todo) => !todo.isCompleted).toList();
+    case TodoListFilter.all:
+      return todos;
+  }
+});
+
+
+
+class BasicTodoPage extends HookConsumerWidget {
+  const BasicTodoPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final todos = ref.watch(filteredTodos);
+    final newTodoController = useTextEditingController();
+
+    return MaterialApp(
+      home: GestureDetector(
+        onTap: () => FocusScope.of(context).unfocus(),
+        child: Scaffold(
+          appBar: AppBar(
+            title: const Text('basic todo'),
+            leading: IconButton(
+              icon: Icon(Icons.arrow_back_ios),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+          ),
+          body: ListView(
+            padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 40),
+            children: [
+              const Title(),
+              TextField(
+                key: addTodoKey,
+                controller: newTodoController,
+                decoration: const InputDecoration(
+                  labelText: 'What needs to be done?',
+                ),
+                onSubmitted: (value) {
+                  ref.read(todoListProvider.notifier).add(value);
+                  newTodoController.clear();
+                },
+              ),
+              const SizedBox(height: 42),
+              const Toolbar(),
+              if (todos.isNotEmpty) const Divider(height: 0),
+              for (var i = 0; i < todos.length; i++) ...[
+                if (i > 0) const Divider(height: 0),
+                Dismissible(
+                  background: Container(
+                    color: Colors.red,
+                    padding: EdgeInsets.only(
+                      right: 10,
+                    ),
+                    alignment: AlignmentDirectional.centerEnd,
+                    child: Icon(
+                      Icons.delete,
+                      color: Colors.white,
+                    ),
+                  ),
+                  direction: DismissDirection.endToStart,
+                  key: ValueKey(todos[i].todoId),
+                  onDismissed: (direction) {
+                    ref.read(todoListProvider.notifier).remove(todos[i]);
+                  },
+                  child: ProviderScope(
+                    overrides: [
+                      _currentTodo.overrideWithValue(todos[i]),
+                    ],
+                    child: const TodoItem(),
+                  ),
+                )
+              ],
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class Toolbar extends HookConsumerWidget {
+  const Toolbar({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final filter = ref.watch(todoListFilter);
+
+    Color? textColorFor(TodoListFilter value) {
+      return filter == value ? Colors.blue : Colors.black;
+    }
+
+    return Material(
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Expanded(
+            child: Text(
+              '${ref.watch(uncompletedTodosCount).toString()} items left',
+              overflow: TextOverflow.ellipsis,
+            ),
+          ),
+          Tooltip(
+            key: allFilterKey,
+            message: 'All todos',
+            child: TextButton(
+              onPressed: () =>
+              ref.read(todoListFilter.notifier).state = TodoListFilter.all,
+              style: ButtonStyle(
+                visualDensity: VisualDensity.compact,
+                foregroundColor:
+                MaterialStateProperty.all(textColorFor(TodoListFilter.all)),
+              ),
+              child: const Text('All'),
+            ),
+          ),
+          Tooltip(
+            key: activeFilterKey,
+            message: 'Only uncompleted todos',
+            child: TextButton(
+              onPressed: () => ref.read(todoListFilter.notifier).state =
+                  TodoListFilter.active,
+              style: ButtonStyle(
+                visualDensity: VisualDensity.compact,
+                foregroundColor: MaterialStateProperty.all(
+                  textColorFor(TodoListFilter.active),
+                ),
+              ),
+              child: const Text('Active'),
+            ),
+          ),
+          Tooltip(
+            key: completedFilterKey,
+            message: 'Only completed todos',
+            child: TextButton(
+              onPressed: () => ref.read(todoListFilter.notifier).state =
+                  TodoListFilter.isCompleted,
+              style: ButtonStyle(
+                visualDensity: VisualDensity.compact,
+                foregroundColor: MaterialStateProperty.all(
+                  textColorFor(TodoListFilter.isCompleted),
+                ),
+              ),
+              child: const Text('Completed'),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class Title extends StatelessWidget {
+  const Title({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Text(
+      'todos',
+      textAlign: TextAlign.center,
+      style: TextStyle(
+        color: Color.fromARGB(38, 47, 47, 247),
+        fontSize: 100,
+        fontWeight: FontWeight.w100,
+        fontFamily: 'Helvetica Neue',
+      ),
+    );
+  }
+}
+
+final _currentTodo = Provider<Todo>((ref) => throw UnimplementedError());
+
+class TodoItem extends HookConsumerWidget {
+  const TodoItem({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final todo = ref.watch(_currentTodo);
+    final itemFocusNode = useFocusNode();
+    final itemIsFocused = useIsFocused(itemFocusNode);
+
+    final textEditingController = useTextEditingController();
+    final textFieldFocusNode = useFocusNode();
+
+    return Material(
+      color: Colors.white,
+      elevation: 6,
+      child: Focus(
+        focusNode: itemFocusNode,
+        onFocusChange: (focused) {
+          if (focused) {
+            textEditingController.text = todo.title;
+          } else {
+            ref
+                .read(todoListProvider.notifier)
+                .edit(todoId: todo.todoId, title: textEditingController.text);
+          }
+        },
+        child: ListTile(
+          onTap: () {
+            itemFocusNode.requestFocus();
+            textFieldFocusNode.requestFocus();
+          },
+          leading: Checkbox(
+            value: todo.isCompleted,
+            onChanged: (value) =>
+                ref.read(todoListProvider.notifier).toggle(todo.todoId),
+          ),
+          title: itemIsFocused
+              ? TextField(
+            autofocus: true,
+            focusNode: textFieldFocusNode,
+            controller: textEditingController,
+          )
+              : Text(todo.title),
+          subtitle: Text(
+            '${todo.createdAt.year.toString()}年' +
+                '${todo.createdAt.month.toString()}月' +
+                '${todo.createdAt.day.toString()}日 ' +
+                '${todo.createdAt.hour.toString()}:' +
+                '${todo.createdAt.minute.toString()}',
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+bool useIsFocused(FocusNode node) {
+  final isFocused = useState(node.hasFocus);
+
+  useEffect(() {
+    void listener() {
+      isFocused.value = node.hasFocus;
+    }
+
+    node.addListener(listener);
+    return () => node.removeListener(listener);
+  }, [node]);
+
+  return isFocused.value;
+}
